add batchEnqueue method to increase performance when enqueueing multiple jobs
When enqueueing a lot of jobs at once the performance can be significantly improved by either using the pipelininig (https://redis.io/topics/pipelining) or transactions (https://redis.io/topics/transactions) feature of Redis.
Since it is easier to implement the batching functionality using transactions, as there is no need to handle situations in which only a subset of jobs has been successfully scheduled, this is what this implementation is using.
Note: Transactions are handled by Redis v1.2 and higher.

Running redis locally, a reduction in execution time from 16000ms to 400 ms for 10k jobs was observed. Considering higher network overhead in real-life scenarios the impact should be even bigger.